Refactored + Customized Sidebar
