--- packages/bratus-app/src/components/NavigationPanel/private/NavigationSection/NavigationSection.sc.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-export const Section = styled.div`
-  display: flex;
-  flex-direction: column;
-  height: auto;
-`;
