+++ packages/bratus-app/src/utils/tokens/constants.js
@@ -2,3 +2,9 @@ export const GraphLabels = {
   leftToRight: 'LR',
   topToBottom: 'TB',
 };
+
+export const defaultOpenKeys = [
+  'search-component',
+  'node-visualization',
+  'navigation-actions',
+];
